First round of merging against server-client and color stuff
